requirements.txt
aiosqlite
APScheduler~=3.9.1
fastapi~=0.79.0
pydantic~=1.9.1
python-dotenv~=0.20.0
python-dateutil~=2.8.2
requests~=2.28.1
SQLAlchemy~=1.4.39
uvicorn~=0.18.2

black[d]~=22.6.0
pandas~=1.4.3
pre-commit
pylint~=2.14.5
coverage
